audio-api/src/requirements-flask_app.txt
boto3==1.24.28
botocore==1.27.59
celery==5.3.5
Flask==2.3.2
Flask_Cors==4.0.0
python-decouple==3.8
Werkzeug==2.3.6
